server: add tests for app
