+++ server/package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "dev": "env NODE_ENV=development nodemon app",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "NODE_ENV=development jest --runInBand --detectOpenHandles",
+    "test:watch": "yarn test --watch"
   },
   "keywords": [],
   "author": "",
@@ -19,6 +20,11 @@
     "nodemailer": "^6.4.6"
   },
   "devDependencies": {
-    "nodemon": "^2.0.3"
+    "nodemon": "^2.0.3",
+    "jest": "^26.1.0",
+    "supertest": "^4.0.2"
+  },
+  "jest": {
+    "testEnvironment": "node"
   }
 }
